# Convert strings with non-ASCII chars into UTF-8