**What does this PR do?**

Convert ASCII-8BIT strings containing non-ASCII chars into UTF-8 strings.

**Motivation**

```
❯❯❯ irb                                                                                                                                               
irb(main):001> require 'ffi'                                                                                                                          
=> true                                                                                                                                               
irb(main):002> str = 'UTF-8 string with some non-ASCII: é à ö'                                                                                        
=> "UTF-8 string with some non-ASCII: é à ö"                                                                                                          
irb(main):003> ptr = FFI::MemoryPointer.new(:char, str.bytesize + 1)                                                                                  
=> #<FFI::MemoryPointer address=0x0000600002464d40 size=43>                                                                                           
irb(main):004> ptr.put_string(0, str)                                                                                                                 
=> #<FFI::MemoryPointer address=0x0000600002464d40 size=43>                                                                                           
irb(main):005> new_str = ptr.read_bytes(str.bytesize)                                                                                                 
=> "UTF-8 string with some non-ASCII: \xC3\xA9 \xC3\xA0 \xC3\xB6"                                                                                     
irb(main):006> require 'json'                                                                                                                         
=> true                                                                                                                                               
irb(main):007> JSON.dump(str)                                                                                                                         
=> "\"UTF-8 string with some non-ASCII: é à ö\""                                                                                                      
irb(main):008> JSON.dump(new_str)                                                                                                                     
/Users/sergey.fedorov/.asdf/installs/ruby/3.3.5/lib/ruby/gems/3.3.0/gems/json-2.10.2/lib/json/common.rb:849: warning: JSON.generate: UTF-8 string pass
ed as BINARY, this will raise an encoding error in json 3.0                                                                                           
=> "\"UTF-8 string with some non-ASCII: é à ö\""                                                                                                      
irb(main):009>
```

**Additional Notes**

Considering benchmark with different approaches a clear winner is conditional conversion of string containing non-ASCII chars only

```
❯❯❯ ruby benchmark_string_encoding.rb                                                                                                                                                                                                                                       
Running benchmarks...                                                                                                                                                                                                                                                       
Ruby version: 3.3.5                                                                                                                                                                                                                                                         
GC disabled during benchmark                                                                                                                                                                                                                                                
ruby 3.3.5 (2024-09-03 revision ef084cc8f4) [arm64-darwin23]                                                                                                                                                                                                                
Warming up --------------------------------------                                                                                                                                                                                                                           
read_bytes (baseline)                                                                                                                                                                                                                                                       
                       144.109k i/100ms                                                                                                                                                                                                                                     
read_bytes + force_encoding                                                                                                                                                                                                                                                 
                        75.323k i/100ms                                                                                                                                                                                                                                     
conditional (ascii_only?)                                                                                                                                                                                                                                                   
                       288.779k i/100ms                                                                                                                                                                                                                                     
 conditional (regex)   133.258k i/100ms                                                                                                                                                                                                                                     
  conditional (loop)    26.668k i/100ms                                                                                                                                                                                                                                     
 read_bytes + encode    44.351k i/100ms                                                                                                                                                                                                                                     
get_bytes + force_encoding                                                                                                                                                                                                                                                  
                       120.684k i/100ms                                                                                                                                                                                                                                     
String.new with encoding                                                                                                                                                                                                                                                    
                         9.967k i/100ms                                                                                                                                                                                                                                     
Calculating -------------------------------------                                                                                                                                                                                                                           
read_bytes (baseline)                                                                                                                                                                                                                                                       
                          2.221M (±80.9%) i/s  (450.27 ns/i) -     11.241M in  15.009524s                                                                                                                                                                                   
read_bytes + force_encoding                                                                                                                                                                                                                                                 
                        996.962k (±90.0%) i/s    (1.00 μs/i) -      7.532M in  15.178095s                                                                                                                                                                                   
conditional (ascii_only?)                                                                                                                                                                                                                                                   
                          2.008M (±56.6%) i/s  (497.92 ns/i) -     12.995M in  16.191018s                                                                                                                                                                                   
 conditional (regex)    853.657k (±64.4%) i/s    (1.17 μs/i) -      6.396M in  15.665583s                                                                                                                                                                                   
  conditional (loop)    277.538k (± 1.1%) i/s    (3.60 μs/i) -      4.187M in  15.087528s                                                                                                                                                                                   
 read_bytes + encode    616.996k (±96.2%) i/s    (1.62 μs/i) -      4.213M in  15.250070s                                                                                                                                                                                   
get_bytes + force_encoding                                                                                                                                                                                                                                                  
                          1.363M (±69.9%) i/s  (733.49 ns/i) -      9.172M in  15.177075s                                                                                                                                                                                   
String.new with encoding                                                                                                                                                                                                                                                    
                        645.687k (±103.9%) i/s    (1.55 μs/i) -      1.694M in  15.183036s                                                                                                                                                                                  
                                                                                          
```

<details><summary>benchmark_string_encoding.rb</summary>
<p>

```ruby
require 'benchmark/ips'
require 'ffi'
require 'stringio'

# Create a mix of simple test strings
sample_strings = [
  'Hello, world!',
  'UTF-8 string with some non-ASCII: é à ö',
  'Short',
  'A somewhat longer string that has more content'
]

class MemoryExample
  extend FFI::Library

  # Creates a memory pointer with the content of the string for testing
  def self.create_memory_pointer(str)
    ptr = FFI::MemoryPointer.new(:char, str.bytesize + 1)
    ptr.put_string(0, str)
    [ptr, str.bytesize]
  end

  # Current approach - read_bytes
  def self.read_with_read_bytes(ptr, size)
    ptr.read_bytes(size)
  end

  # Alternative - read_string
  def self.read_with_read_string(ptr, size)
    ptr.read_string(size)
  end

  # Alternative - get_bytes
  def self.read_with_get_bytes(ptr, size)
    ptr.get_bytes(0, size)
  end

  # Alternative - get_string
  def self.read_with_get_string(ptr, size)
    ptr.get_string(0, size)
  end

  # StringIO approach
  def self.read_with_stringio(ptr, size)
    bytes = ptr.read_bytes(size)
    io = StringIO.new(bytes)
    io.string.force_encoding('UTF-8')
  end

  # StringIO approach 2 - fixed
  def self.read_with_stringio_direct(ptr, size)
    io = StringIO.new(ptr.read_bytes(size))
    io.set_encoding('UTF-8')
    io.string
  end

  # Direct encode approach
  def self.read_with_direct_encode(ptr, size)
    ptr.read_bytes(size).encode('UTF-8', 'UTF-8')
  end

  # Conditional force_encoding - original (slow)
  def self.read_with_conditional_encoding_slow(ptr, size)
    bytes = ptr.read_bytes(size)
    # Check if string contains any non-ASCII bytes (values > 127)
    if bytes.bytes.any? { |b| b > 127 }
      bytes.force_encoding('UTF-8')
    else
      bytes
    end
  end

  # Conditional force_encoding - optimized
  def self.read_with_conditional_encoding(ptr, size)
    bytes = ptr.read_bytes(size)
    # Check for non-ASCII bytes more efficiently without creating arrays
    has_non_ascii = false
    i = 0
    while i < bytes.bytesize
      if bytes.getbyte(i) > 127
        has_non_ascii = true
        break
      end
      i += 1
    end

    if has_non_ascii
      bytes.force_encoding('UTF-8')
    else
      bytes
    end
  end

  # Conditional force_encoding - optimized variant 2
  def self.read_with_conditional_encoding_v2(ptr, size)
    bytes = ptr.read_bytes(size)

    # Check for non-ASCII bytes using regex
    if bytes =~ /[\x80-\xFF]/n
      bytes.force_encoding('UTF-8')
    else
      bytes
    end
  end

  # Conditional force_encoding - optimized variant 3 (fastest check)
  def self.read_with_conditional_encoding_v3(ptr, size)
    bytes = ptr.read_bytes(size)

    # Check for non-ASCII bytes using String#ascii_only?
    # This is a built-in method that should be faster than manual checks
    if bytes.ascii_only?
      bytes
    else
      bytes.force_encoding('UTF-8')
    end
  end
end

# Prepare test data
test_data = []
sample_strings.each do |str|
  ptr, size = MemoryExample.create_memory_pointer(str)
  test_data << [ptr, size, str]
end

puts 'Running benchmarks...'
puts "Ruby version: #{RUBY_VERSION}"

# Disable GC during benchmarking for more accurate results
GC.disable
puts 'GC disabled during benchmark'

Benchmark.ips do |x|
  # Run benchmark for longer time with more warmup
  x.config(time: 15, warmup: 5)

  # Baseline (no encoding change) - Current implementation
  x.report('read_bytes (baseline)') do
    test_data.each do |ptr, size, _|
      MemoryExample.read_with_read_bytes(ptr, size)
    end
  end

  # With force_encoding
  x.report('read_bytes + force_encoding') do
    test_data.each do |ptr, size, _|
      MemoryExample.read_with_read_bytes(ptr, size).force_encoding('UTF-8')
    end
  end

  # Conditional encoding - ascii_only? check
  x.report('conditional (ascii_only?)') do
    test_data.each do |ptr, size, _|
      MemoryExample.read_with_conditional_encoding_v3(ptr, size)
    end
  end

  # Conditional encoding - regex version
  x.report('conditional (regex)') do
    test_data.each do |ptr, size, _|
      MemoryExample.read_with_conditional_encoding_v2(ptr, size)
    end
  end

  # Conditional encoding - optimized loop
  x.report('conditional (loop)') do
    test_data.each do |ptr, size, _|
      MemoryExample.read_with_conditional_encoding(ptr, size)
    end
  end

  # Direct encode approach
  x.report('read_bytes + encode') do
    test_data.each do |ptr, size, _|
      MemoryExample.read_with_direct_encode(ptr, size)
    end
  end

  x.report('get_bytes + force_encoding') do
    test_data.each do |ptr, size, _|
      MemoryExample.read_with_get_bytes(ptr, size).force_encoding('UTF-8')
    end
  end

  x.report('String.new with encoding') do
    test_data.each do |ptr, size, _|
      String.new(MemoryExample.read_with_read_bytes(ptr, size), encoding: 'UTF-8')
    end
  end

  x.compare!
end

# Re-enable GC after benchmarking
GC.enable
puts 'GC re-enabled'

# Show simple verification for one string
puts "\nVerification:"
ptr, size, original = test_data[1] # String with some non-ASCII chars

puts "Original string: #{original.inspect}"
puts "Encoding: #{original.encoding}"

puts "\nResults:"

binary = MemoryExample.read_with_read_bytes(ptr, size)
puts 'read_bytes (baseline):'
puts "  Result: #{binary.inspect}"
puts "  Encoding: #{binary.encoding}"

result = binary.force_encoding('UTF-8')
puts "\nread_bytes + force_encoding:"
puts "  Result: #{result.inspect}"
puts "  Encoding: #{result.encoding}"
puts "  Matches original: #{result == original}"

result = MemoryExample.read_with_conditional_encoding_v3(ptr, size)
puts "\nconditional (ascii_only?):"
puts "  Result: #{result.inspect}"
puts "  Encoding: #{result.encoding}"
puts "  Matches original: #{result == original}"

result = MemoryExample.read_with_conditional_encoding_v2(ptr, size)
puts "\nconditional (regex):"
puts "  Result: #{result.inspect}"
puts "  Encoding: #{result.encoding}"
puts "  Matches original: #{result == original}"

# Also verify with ASCII-only string
ptr, size, original = test_data[0] # ASCII-only string
result = MemoryExample.read_with_conditional_encoding_v3(ptr, size)
puts "\nconditional (ascii_only?) with ASCII-only string:"
puts "  Result: #{result.inspect}"
puts "  Encoding: #{result.encoding}"
puts "  Matches original: #{result == original}"

result = MemoryExample.read_with_conditional_encoding_v2(ptr, size)
puts "\nconditional (regex) with ASCII-only string:"
puts "  Result: #{result.inspect}"
puts "  Encoding: #{result.encoding}"
puts "  Matches original: #{result == original}"
```

</p>
</details> 

**How to test the change?**
<!--
Describe here how the change can be validated.
You are strongly encouraged to provide automated tests for this PR.
If this change cannot be feasibly tested, please explain why,
unless the change does not modify code (e.g. only modifies docs, comments).
-->